distribution/tools/perf/monitor_dashboard/requirements.txt
# Performance Monitoring Dashboard - Project Requirements

# Core System Monitoring
psutil>=5.9.0

# Web Framework
flask>=2.3.0
flask-cors>=4.0.0
flask-socketio>=5.3.0

# Database
sqlalchemy>=2.0.0

# Data Processing & Analysis
requests>=2.31.0
python-dateutil>=2.8.0
numpy>=1.24.0
pandas>=2.0.0

# Visualization & Reporting
matplotlib>=3.7.0
plotly>=5.15.0

# Task Scheduling & Background Jobs
schedule>=1.2.0
apscheduler>=3.10.0

# Progress & Monitoring
tqdm>=4.65.0
colorama>=0.4.6

# File System Monitoring
watchdog>=3.0.0

# Configuration & Templates
pyyaml>=6.0
jinja2>=3.1.0

# Optional: Additional Features
# redis>=4.5.0  # For caching (optional)
# prometheus-client>=0.16.0  # For metrics export (optional)

# Development Dependencies (optional)
# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=23.0.0
# flake8>=6.0.0